Rubocop + enable Redis on CI
