+++ test/multiverse/suites/sidekiq_ignore_retry_errors_enabled/Envfile
@@ -5,7 +5,8 @@
 SIDEKIQ_VERSIONS = [
   [nil, 3.2],
   ['7.3.9', 2.7],
-  ['6.4.0', 2.5]
+  ['6.4.0', 2.5],
+  ['5.0.3', 2.4, 2.5]
 ]
 
 def gem_list(sidekiq_version = nil)
