+++ test/multiverse/suites/sidekiq_ignore_retry_errors_enabled/sidekiq_ignore_retry_errors_enabled_test.rb
@@ -4,29 +4,13 @@
 
 require_relative 'sidekiq_test_helpers'
 
-class SidekiqIgnoreRetryErrorsTest < Minitest::Test
+# On startup, Sidekiq instrumentation registers error and death handlers
+# based on the value of the 'sidekiq.ignore_retry_errors'. Because of this,
+# we need to have separate enabled/disabled test suites to test both cases.
+class SidekiqIgnoreRetryErrorEnabledTest < Minitest::Test
   include SidekiqTestHelpers
 
-  def test_sidekiq_ignore_retry_errors_default_value
-    refute NewRelic::Agent.config[:sidekiq.ignore_retry_errors],
-      'Expected sidekiq.ignore_retry_errors default to be false'
-  end
-
-  def test_sidekiq_ignore_retry_errors_configuration_can_be_set_to_true
-    with_config(sidekiq.ignore_retry_errors: true) do
-      assert NewRelic::Agent.config[:sidekiq.ignore_retry_errors],
-        'Expected sidekiq.ignore_retry_errors to be true when configured'
-    end
-  end
-
-  def test_sidekiq_ignore_retry_errors_configuration_can_be_set_to_false
-    with_config(sidekiq.ignore_retry_errors: false) do
-      refute NewRelic::Agent.config[:sidekiq.ignore_retry_errors],
-        'Expected sidekiq.ignore_retry_errors to be false when configured'
-    end
-  end
-
-  def test_error_handlers_registered_when_sidekiq_ignore_retry_errors_is_false
+  def test_error_handlers_not_registered_when_sidekiq_ignore_retry_errors_is_true
     # TODO: MAJOR VERSION - remove this when Sidekiq v5 is no longer supported
     skip 'Test requires Sidekiq v6+' unless Sidekiq::VERSION.split('.').first.to_i >= 6
 
@@ -46,11 +30,11 @@ class SidekiqIgnoreRetryErrorsTest < Minitest::Test
       handler.is_a?(Proc) && handler.source_location&.first&.include?('newrelic')
     end
 
-    assert nr_error_handler_found,
-      'Expected NewRelic error_handler to be registered when sidekiq.ignore_retry_errors is false'
+    refute nr_error_handler_found,
+      'Expected NewRelic error_handler to NOT be registered when sidekiq.ignore_retry_errors is true'
   end
 
-  def test_death_handlers_not_registered_when_sidekiq_ignore_retry_errors_is_false
+  def test_death_handlers_registered_when_sidekiq_ignore_retry_errors_is_true
     # TODO: MAJOR VERSION - remove this when Sidekiq v5 is no longer supported
     skip 'Test requires Sidekiq v6+' unless Sidekiq::VERSION.split('.').first.to_i >= 6
 
@@ -70,7 +54,15 @@ class SidekiqIgnoreRetryErrorsTest < Minitest::Test
       handler.is_a?(Proc) && handler.source_location&.first&.include?('newrelic')
     end
 
-    refute nr_death_handler_found,
-      'Expected NewRelic death_handler to NOT be registered when sidekiq.ignore_retry_errors is false'
+    assert nr_death_handler_found,
+      'Expected NewRelic death_handler to be registered when sidekiq.ignore_retry_errors is true'
+  end
+
+  def test_basic_job_execution_still_works
+    segment = run_job
+
+    assert_predicate segment, :finished?
+    assert_predicate segment, :record_metrics?
+    assert_kind_of Float, segment.duration
   end
 end
