+++ test/multiverse/suites/sidekiq_ignore_retry_errors_enabled/sidekiq_test_helpers.rb
